Move profiler code to a dedicated 'profiler' library.
