+++ src/libs/profiler/profiler/CMakeFinds.cmake.in
@@ -1,0 +1,2 @@
+@find_package@(Snacman CONFIG @REQUIRED@ utilities)
+@find_package@(Graphics CONFIG @REQUIRED@ renderer)
