+++ src/libs/profiler/profiler/providers/ProviderWindows.cpp
