+++ src/libs/profiler/profiler/providers/ProviderWindows.h
